Has been deleted some unnecessary files.

+++ IndividualCollectionsWeb/Controllers/UserController.cs
@@ -20,15 +20,12 @@ namespace IndividualCollectionsWeb.Controllers
         public ActionResult Index()
         {
             ApplicationUser user = UserManager.FindById(User.Identity.GetUserId());
+            IQueryable<Collection> collections = null;
             if (user != null)
             {
-                var collections = db.Collections.Include(c => c.Theme).Where(c => c.ApplicationUserId == user.Id);
-                return View(collections.ToList());
+                collections = db.Collections.Include(c => c.Theme).Where(c => c.ApplicationUserId == user.Id);
             }
-            else
-            {
-                return View();
-            }
+            return View(collections.ToList());
         }
 
         // GET: User/DetailsCollection/5
@@ -124,9 +121,12 @@ namespace IndividualCollectionsWeb.Controllers
                 {
                     if (!user.Blocked)
                     {
-                        db.Entry(collection).State = EntityState.Modified;
-                        db.SaveChanges();
-                        return RedirectToAction("Index");
+                        if (collection.ApplicationUserId == user.Id)
+                        {
+                            db.Entry(collection).State = EntityState.Modified;
+                            db.SaveChanges();
+                            return RedirectToAction("Index");
+                        }
                     }
                     else
                     {
@@ -173,8 +173,11 @@ namespace IndividualCollectionsWeb.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
             }
             Collection collection = db.Collections.Find(id);
-            db.Collections.Remove(collection);
-            db.SaveChanges();
+            if (user != null && collection.ApplicationUserId == user.Id)
+            {
+                db.Collections.Remove(collection);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
 
@@ -216,18 +219,129 @@ namespace IndividualCollectionsWeb.Controllers
         public ActionResult CreateItem([Bind(Include = "Id,CollectionId,Title,Tags,AddingTime,Numerical1,Numerical2,Numerical3,String1,String2,String3,Text1,Text2,Text3,Date1,Date2,Date3,Boolean1,Boolean2,Boolean3")] Item item)
         {
             ApplicationUser user = UserManager.FindById(User.Identity.GetUserId());
+            if (ModelState.IsValid)
+            {
+                if (user != null)
+                {
+                    if (!user.Blocked)
+                    {
+                        item.AddingTime = DateTime.Now;
+                        db.Items.Add(item);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("", "Пользователь заблокирован");
+                    }
+                }
+                else
+                {
+                    ModelState.AddModelError("", "Пользователь не найден");
+                }
+            }
+            ViewBag.CollectionId = new SelectList(db.Collections, "Id", "Title", item.CollectionId);
+            return View(item);
+        }
+
+        // GET: User/DetailsItem/5
+        public ActionResult DetailsItem(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Item item = db.Items.Find(id);
+            if (item == null)
+            {
+                return HttpNotFound();
+            }
+            return View(item);
+        }
+
+        // GET: User/EditItem/5
+        public ActionResult EditItem(int? id)
+        {
+            ApplicationUser user = UserManager.FindById(User.Identity.GetUserId());
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Item item = db.Items.Find(id);
+            if (item == null)
+            {
+                return HttpNotFound();
+            }
+            ViewBag.CollectionId = new SelectList(db.Collections.Where(c => c.ApplicationUserId == user.Id), "Id", "Title", item.CollectionId);
+            return View(item);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditItem([Bind(Include = "Id,CollectionId,Title,Tags,AddingTime,Numerical1,Numerical2,Numerical3,String1,String2,String3,Text1,Text2,Text3,Date1,Date2,Date3,Boolean1,Boolean2,Boolean3")] Item item)
+        {
+            ApplicationUser user = UserManager.FindById(User.Identity.GetUserId());
+            if (ModelState.IsValid)
+            {
+                if (user != null)
+                {
+                    if (!user.Blocked)
+                    {
+                        Collection collection = db.Collections.Find(item.CollectionId);
+                        if (user != null && collection.ApplicationUserId == user.Id)
+                        {
+                            db.Entry(item).State = EntityState.Modified;
+                            db.SaveChanges();
+                        }
+                        return RedirectToAction("Index");
+                    }
+                    else
+                    {
+                        ModelState.AddModelError("", "Пользователь заблокирован");
+                    }
+                }
+                else
+                {
+                    ModelState.AddModelError("", "Пользователь не найден");
+                }
+            }
+            ViewBag.CollectionId = new SelectList(db.Collections, "Id", "Title", item.CollectionId);
+            return View(item);
+        }
+
+        // GET: User/DeleteItem/5
+        public ActionResult DeleteItem(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+            Item item = db.Items.Find(id);
+            if (item == null)
+            {
+                return HttpNotFound();
+            }
+            return View(item);
+        }
+
+        // POST: User/DeleteItem/5
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteItem(int id)
+        {
+            ApplicationUser user = UserManager.FindById(User.Identity.GetUserId());
             if (user != null && user.Blocked)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
             }
-            if (ModelState.IsValid)
+            Item item = db.Items.Find(id);
+            Collection collection = db.Collections.Find(item.CollectionId);
+            if (user != null && collection.ApplicationUserId == user.Id)
             {
-                db.Items.Add(item);
+                db.Items.Remove(item);
                 db.SaveChanges();
-                return RedirectToAction("Index");
             }
-            ViewBag.CollectionId = new SelectList(db.Collections, "Id", "Title", item.CollectionId);
-            return View(item);
+            return RedirectToAction("Index");
         }
 
         private ApplicationUserManager UserManager
